tutorial2_各Widget, Singleton（のつもり）のクラスを別ファイルにわけた
