+++ flutter_tutorial/lib/youtube/youtube_app_colors.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+// 画面やWidgetの背景色を定義したもの
+class YoutubeAppColors {
+  static const rgb322933 = Color.fromARGB(255, 32, 29, 33);
+  static const rgb221923 = Color.fromARGB(255, 22, 19, 23);
+}
